validação de usuarios e administradores

+++ client/app/admin/usuarios/page.js
@@ -1,13 +1,18 @@
 'use client'
 import MontaTabela from "@/app/components/montaTabela";
 import Link from "next/link";
-import { useEffect, useRef, useState } from "react"
+import { useEffect, useRef, useState, useContext } from "react"
+import UserContext from "../../context/userContext.js"
 
 export default function usuariosAdmin() {
 
     let msgRef = useRef(null)
 
     let [listaUsuarios, setListaUsuarios] = useState([]);
+    //let [usuarioLogado, setUsuarioLogado] = useState([]);
+    let usuarioLogado
+    const {user, setUser} = useContext(UserContext)
+
     useEffect((e) => {
         carregarUsuarios();
     }, [])
@@ -26,45 +31,60 @@ export default function usuariosAdmin() {
         })
     }
 
-    function excluirUsuario(id) {
+    async function excluirUsuario(id) {
 
         msgRef.current.className = ''
         msgRef.current.innerHTML = ''
-        if(confirm("Tem certeza que deseja excluir este usuario?")) {
-            if(id > 0) {
-                let ok = false;
-                fetch(`http://localhost:5000/usuarios/excluir/${id}`, {
-                    mode: 'cors',
-                    credentials: 'include',
-                    method: "DELETE",
-                })
-                .then(r=> {
-                    ok = r.status == 200;
-                    return r.json();
-                })
-                .then(r=> {
-                    if(ok) {
-                        msgRef.current.className = "msgSucess";
-                        msgRef.current.innerHTML = r.msg;
-                        carregarUsuarios();
 
-                        setTimeout(() => {
-                            msgRef.current.innerHTML = '';
-                            msgRef.current.className = '';
-                        }, 5000);
-                    }
-                    else{
-                        msgRef.current.className = "msgError";
-                        msgRef.current.innerHTML = r.msg;
+        let usuario = await fetch(`http://localhost:5000/usuarios/obter/${id}`,{
+            mode: 'cors',
+            credentials: 'include',
+            method: "GET",
+        });
+        let dados = await usuario.json();
+        usuarioLogado = dados.usuId
 
-                        setTimeout(() => {
-                            msgRef.current.innerHTML = '';
-                            msgRef.current.className = '';
-                        }, 5000);
-                    }
-                })
+        if(user.usuId != usuarioLogado){
+            if(confirm("Tem certeza que deseja excluir este usuario?")) {
+                if(id > 0) {
+                    let ok = false;
+                    fetch(`http://localhost:5000/usuarios/excluir/${id}`, {
+                        mode: 'cors',
+                        credentials: 'include',
+                        method: "DELETE",
+                    })
+                    .then(r=> {
+                        ok = r.status == 200;
+                        return r.json();
+                    })
+                    .then(r=> {
+                        if(ok) {
+                            msgRef.current.className = "msgSucess";
+                            msgRef.current.innerHTML = r.msg;
+                            carregarUsuarios();
+
+                            setTimeout(() => {
+                                msgRef.current.innerHTML = '';
+                                msgRef.current.className = '';
+                            }, 5000);
+                        }
+                        else{
+                            msgRef.current.className = "msgError";
+                            msgRef.current.innerHTML = r.msg;
+
+                            setTimeout(() => {
+                                msgRef.current.innerHTML = '';
+                                msgRef.current.className = '';
+                            }, 5000);
+                        }
+                    })
+                }
             }
         }
+        else{
+            msgRef.current.className = "msgError";
+            msgRef.current.innerHTML = 'Não é possivel apagar o proprio usuário';
+        }    
     }
 
     return (
